+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-07-23 10:49:24
-// Commit: feat(perf): implement responsive perf view component
+// Updated at: 2026-07-23 16:40:36
+// Commit: fix(perf): prevent double tap duplicate action in perf
 
 class perfComponentState {
-  final String commitRef = '3052';
-  final String lastMessage = 'feat(perf): implement responsive perf view component';
-  final DateTime timestamp = DateTime.parse('2026-07-23T10:49:24+05:00');
+  final String commitRef = '3067';
+  final String lastMessage = 'fix(perf): prevent double tap duplicate action in perf';
+  final DateTime timestamp = DateTime.parse('2026-07-23T16:40:36+05:00');
 }